+++ CHANGELOG.md
@@ -10,7 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   PID to debug log.
 
 ### Changed
-*   Write example log messages to examples\log\example.log instead of error.log.
+*   Update example implementation for Hire Purchase direct debit payment type (FlexiPay® Rate).
+*   Write log messages from examples to examples\log\example.log instead of error.log.
+
+### Fix
+*   Ensure locale strings comply to standard.
 
 ## [1.2.6.0][1.2.6.0]