+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 *   Update example implementation for Hire Purchase direct debit payment type (FlexiPay® Rate).
 *   Write log messages from examples to examples\log\example.log instead of error.log.
+*   Several minor changes.
 
 ### Fix
 *   Ensure locale strings comply to standard.