+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   Http headers to debug log.
 *   PID to debug log.
 *   PHP7.4 support.
+*   Handles for automatic testing to examples.
 
 ### Changed
 *   Update example implementation for Hire Purchase direct debit payment type (FlexiPay® Rate).